+++ backend/src/main/java/teamb/w4e/connectors/externaldto/ReservationRequestDTO.java
@@ -1,4 +1,4 @@
 package teamb.w4e.connectors.externaldto;
 
-public record ReservationRequestDTO(String name, String activity, Long duration) {
+public record ReservationRequestDTO(String name, String activity, int duration) {
 }
