+++ backend/src/main/java/teamb/w4e/interfaces/SkiPass.java
@@ -3,5 +3,5 @@ package teamb.w4e.interfaces;
 import java.util.Optional;
 
 public interface SkiPass {
-    Optional<String> reserve(String name, String activity, Long duration);
+    Optional<String> reserve(String name, String activity, int duration);
 }
